refactor: Move workflow preparation code out of chat hub service
